+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 0.7.0 (beta)
 * Implement generic DBus signal handler and related toolbox
 * Some conditions can be activated from the command line
-* File and directory change notifications
+* Conditions based on file and directory changes 
 * Bug fixes
 
 ### Compatibility break